pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wechat-automation"
version = "0.1.0"
description = "WeChat document automation — auto-ingest vendor files to Firestore"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-firestore>=2.16",
    "google-cloud-storage>=2.14",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "watchdog>=4.0",
    "rapidfuzz>=3.6",
    "python-magic-bin>=0.4;sys_platform=='win32'",
]

[project.optional-dependencies]
extractors = [
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "pdfplumber>=0.10",
    "google-generativeai>=0.5",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
